# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ursol CAST API - ERP System",
        "description": "API completa de ERP para empresas costarricenses. Sistema integral de gestión empresarial con soporte para contabilidad, facturación electrónica, inventario, compras, ventas, recursos humanos y más.",
        "contact": {
            "name": "Sistemas Ursol",
            "email": "soporte@ursol.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo Local"
        },
        {
            "url": "https://api.ursol.com",
            "description": "Servidor de Producción"
        }
    ],
    "paths": {
        "/api/almacenes": {
            "get": {
                "tags": [
                    "Almacenes"
                ],
                "summary": "Listar almacenes",
                "description": "Lista almacenes/bodegas con filtros por empresa y sucursal",
                "operationId": "f61f06f5147ce97df7be3af71d61d424",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "empresa_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sucursal_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "activos",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Almacen"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Almacenes"
                ],
                "summary": "Crear almacén",
                "description": "Crea un almacén/bodega. Si es principal, desmarca otros de la sucursal",
                "operationId": "bf31688475550ccb73ed2d6a0dc10d7a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "empresa_id",
                                    "sucursal_id",
                                    "nombre"
                                ],
                                "properties": {
                                    "empresa_id": {
                                        "type": "integer"
                                    },
                                    "sucursal_id": {
                                        "type": "integer"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Bodega Principal"
                                    },
                                    "codigo": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "ubicacion": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "es_principal": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Almacén creado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/almacenes/{id}": {
            "get": {
                "tags": [
                    "Almacenes"
                ],
                "summary": "Obtener almacén",
                "description": "Display the specified resource.",
                "operationId": "b6d65b441eea90e6cdaea93eafa89bbb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Almacén encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Almacenes"
                ],
                "summary": "Actualizar almacén",
                "description": "Update the specified resource in storage.",
                "operationId": "7c23829b74611e952cd084d4a308da5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Almacenes"
                ],
                "summary": "Eliminar almacén",
                "description": "Soft delete. No permite eliminar almacén principal",
                "operationId": "c72c27c6c60f10adba5dfc1edba3755b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    },
                    "422": {
                        "description": "No se puede eliminar el principal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Almacen": {
                "title": "Almacén",
                "description": "Almacén o bodega de inventario",
                "required": [
                    "id",
                    "empresa_id",
                    "sucursal_id",
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "empresa_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sucursal_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Bodega Principal"
                    },
                    "codigo": {
                        "type": "string",
                        "example": "ALM-001",
                        "nullable": true
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Almacén de productos terminados",
                        "nullable": true
                    },
                    "ubicacion": {
                        "type": "string",
                        "example": "Edificio A, Planta Baja",
                        "nullable": true
                    },
                    "es_principal": {
                        "type": "boolean",
                        "example": true
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "eliminado": {
                        "type": "boolean",
                        "example": false
                    },
                    "creado_en": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "actualizado_en": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "empresa": {
                        "$ref": "#/components/schemas/Empresa"
                    },
                    "sucursal": {
                        "$ref": "#/components/schemas/Sucursal"
                    }
                },
                "type": "object"
            },
...
            "Empresa": {
                "title": "Empresa",
                "description": "Empresa del sistema (tenant)",
                "required": [
                    "id",
                    "nombre",
                    "identificacion"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Sistemas Ursol S.A."
                    },
                    "nombre_comercial": {
                        "type": "string",
                        "example": "Ursol",
                        "nullable": true
                    },
                    "identificacion": {
                        "type": "string",
                        "example": "3-101-123456"
                    },
                    "tipo_identificacion": {
                        "type": "string",
                        "enum": [
                            "CEDULA_FISICA",
                            "CEDULA_JURIDICA",
                            "DIMEX",
                            "NITE"
                        ],
                        "example": "CEDULA_JURIDICA"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "2222-3333",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "info@ursol.com",
                        "nullable": true
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "Sucursal": {
                "title": "Sucursal",
                "description": "Sucursal de una empresa",
                "required": [
                    "id",
                    "empresa_id",
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "empresa_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Sucursal Centro"
                    },
                    "codigo": {
                        "type": "string",
                        "example": "SUC-001",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "2222-3333",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "centro@empresa.com",
                        "nullable": true
                    },
                    "direccion": {
                        "type": "string",
                        "example": "San José, Centro, Avenida Central",
                        "nullable": true
                    },
                    "provincia": {
                        "type": "string",
                        "example": "San José",
                        "nullable": true
                    },
                    "canton": {
                        "type": "string",
                        "example": "San José",
                        "nullable": true
                    },
                    "distrito": {
                        "type": "string",
                        "example": "Carmen",
                        "nullable": true
                    },
                    "es_principal": {
                        "type": "boolean",
                        "example": true
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "eliminado": {
                        "type": "boolean",
                        "example": false
                    },
                    "creado_en": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "actualizado_en": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "empresa": {
                        "$ref": "#/components/schemas/Empresa"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token Authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}